pyproject.toml
# ==== pytest ====
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--ds=config.settings.test --reuse-db --import-mode=importlib"
python_files = ["tests.py", "test_*.py"]
norecursedirs = [
    "static",
    "staticfiles",
    "node_modules",
    "venv",
    ".venv",
    "env",
    "build",
    "dist",
    "docs",
    "compose",
    "docker",
    "locale",
    "__pycache__",
]

testpaths = ["tests", "cosray_backend"]

# ==== Coverage ====
[tool.coverage.run]
include = ["cosray_backend/**"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    # local virtualenvs
    "**/.venv/**",
    "**/venv/**",
    "**/env/**",
    ".venv/**",
    "venv/**",
    # site-packages / installed distributions
    "**/site-packages/**",
    "**/*dist-info/**",
    # node / frontend dependencies
    "**/node_modules/**",
    "node_modules/**",
    # built/static assets
    "**/static/**",
    "staticfiles/**",
]
plugins = ["django_coverage_plugin"]

# ==== mypy ====
[tool.mypy]
python_version = "3.13"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[[tool.mypy.overrides]]
# Django migrations should not produce any errors:
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.test"

# ==== djLint ====
[tool.djlint]
blank_line_after_tag = "load,extends"
close_void_tags = true
format_css = true
format_js = true
# TODO: remove T002 when fixed https://github.com/djlint/djLint/issues/687
ignore = "H006,H030,H031,T002"
include = "H017,H035"
indent = 2
max_line_length = 119
profile = "django"

[tool.djlint.css]
indent_size = 2

[tool.djlint.js]
indent_size = 2


[tool.black]
line-length = 119
skip-string-normalization = true

[tool.isort]
profile = "black"

[dependency-groups]
dev = [
    "black>=25.9.0",
    "coverage==7.11.3",
    "django-coverage-plugin==3.2.0",
    "django-debug-toolbar==6.1.0",
    "django-extensions==4.1",
    "django-stubs[compatible-mypy]==5.2.7",
    "djangorestframework-stubs==3.16.5",
    "djlint==1.36.4",
    "factory-boy==3.3.3",
    "ipdb==0.13.13",
    "mypy==1.18.2",
    "pre-commit==4.4.0",
    "psycopg[c]==3.2.12",
    "pytest==9.0.0",
    "pytest-django==4.11.1",
    "pytest-sugar==1.1.1",
    "sphinx==8.2.3",
    "sphinx-autobuild==2025.8.25",
    "watchfiles==1.1.1",
    "werkzeug[watchdog]==3.1.3",
]

[project]
name = "cosray_backend"
version = "0.1.0"
description = "Behold My Awesome Project!"
readme = "README.md"

license = { text = "AGPLv3" }
authors = [{ name = "Grid Group", email = "grid-group@example.com" }]
requires-python = "==3.13.*"
dependencies = [
    "argon2-cffi==25.1.0",
    "black>=25.9.0",
    "django==5.2.8",
    "django-allauth[headless,mfa]==65.13.0",
    "django-anymail==13.1",
    "django-cors-headers==4.9.0",
    "django-environ==0.12.0",
    "django-model-utils==5.0.0",
    "django-redis==6.0.0",
    "djangorestframework==3.16.1",
    "drf-spectacular==0.29.0",
    "gunicorn==23.0.0",
    "hiredis==3.3.0",
    "pillow==12.0.0",
    "psycopg[c]==3.2.12",
    "pytest-cov>=7.0.0",
    "python-slugify==8.0.4",
    "redis==7.0.1",
    "uvicorn-worker==0.4.0",
    "uvicorn[standard]==0.38.0",
]
